Added code for time-series moment-matching
Plots the moment-matched (moments of the entire time series)  to that of the wind speed. This code produces the matched-moment time series of the pressure and ground acceleration envelopes, and wind speed time series as shown in Figure 3b. It can be adapted to reproduce other figures in the paper which incorporate the same approach of the method of moments to match time series, i.e. Figure 1c, 2b & 14.